refactor(mocks): extract users module from mockApi (#991 M3)
